requirements.txt
pydantic-settings
python-dotenv
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.3
python-multipart==0.0.6
pydantic==2.5.3
pydantic_settings
email-validator

# --- Testing ---
pytest==8.2.2
httpx==0.27.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
psycopg2-binary
sqlalchemy_utils
